requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.8
httpx==0.25.1
idna==3.10
iniconfig==2.1.0
motor==3.7.0
packaging==25.0
pluggy==1.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.12.0
pytest==7.4.3
python-dotenv==1.0.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.13.2
uvicorn==0.24.0
